Update: remove ViewController and use LiveVideoViewController instead of it
